Extend range of Qphred
There is no need for Qphred to be represented by int8 as the score value
cannot be below 0, so change the underlying type to byte and peform
appropriate conversions to Qsolexa as needed. Also add unknown value to
Qphred and Qsolexa that return NaN as an error probability. The value
for this matches that chosen by samtools (255).