Revision of bed handler - use new feat types
I'm not completely sure of the formatting approach, but it works and
changes to the implementation should leave the API untouched.

An alternative approach to the structure of the collection of Bed* types
was to use embedding resulting in nesting up to 4 levels deep. I felt
that this would be cumbersome for the user, so we have code duplication
here instead.