Fix apparent regression due to ac979
The regression was due to another incorrect use of phred scores and
float to int rounding. The tests in the above commit highlighted this
error, but this change was not propagated to the dependent packages.

The default Qphred threshold has been increased to 3 as this corresponds
to an E value of ~0.5 which is a pretty reasonable arbitrary place to
choose.