Alter the seq.Sequence interface for SetE? methods
Allow Set and SetE to return an error.

This is not used by the core sequence types, but can be used, for
example as by the contig type used in biogo.examples where it can be
used to indicate inability to perform the set. Alternatives were to
panic or magically generate a sequence to write to. This seems useful
enough generally to warrant the change.

The E?At methods do not need this change as out of range justifies a
panic and qaulity can be used to indicate letter validity.