Error chain handling
This idea was prompted by a post by Rog Peppe and subsequent comment
by Kyle Lemons, so they deserved a significant part of the credit.

The chain handling is minimal but configurable if a user-defined Chain
is created, for example the user-defined type could grab a runtime.PC
on each Link call to give detailed stack information. Similarly,
error filtering can be easily added, although this does need more
thought.

The basic design principles are that you want to know the primary
cause of an error and be able to extract intervening annotation.

The API should be considered experimental at this stage.