Fix NW scoreing for first segment
Previously, we were using a short cut to the scoring matrix that
retained the path resolution and relative score values, but did not
include the initial gap setup. This, and the failure to consider the
possibility that the initial segment is a gap in the backtracking
resulted in a zero score for first segments with a gap.

The change here makes the implementation properly reflect the normal
implementation and fixes that issue. In addition there is code added to
allow rendering of the DP progress.