Enforce gapped alphabet requirement
The change of gap index to 0 requires that the input alphabet has a gap
there, otherwise the indexing of letters is off by one. The previous
solution was to make this true for all provided alphabets - this
solution was brittle and broke the pals packages.

What we do now is check that the first letter of the alphabet is indeed
the alphabet's gap letter and fail if this is not true.