Allow fastq parser to handle zero-length sequences
This change allows the fastq reader to handle a fairly wide variety of
FASTQ format abuses (probably more than is reasonable to expect), but it
is not currently guaranteed to cope with cases where a zero-length
sequence line has been stripped completely (for example where an
overzealous application of `grep -v ^$' has been applied); this may not
be fixed.

Fixes issue 7.