Fix offset slice error
This bug affects PALS code too. It is possible for the low and high
range indexes when both further left than the origin of the vector or
both further right of the vector to result in an index to the at method
of -1 or len(vector). This skirts the issue.